finish all bugs and order them
This now includes bugs 0 - 5 in comments, with explanations
as to what goes wrong. The code outside comments is the
working code.

Bugs 0 and 5 are for the article but not for the talk, since it would be
too much for a 30 min talk.

Bug 4, as is, still has a bug if one clicks the jitter button twice without
moving in between. This wouldn't block a successful demo, so it's
not solved, yet.